+++ cfdm/test/test_Field.py
@@ -609,6 +609,33 @@ class FieldTest(unittest.TestCase):
                     self.assertTrue(f.equals(c, verbose=3), message)
         # --- End: for
 
+    def test_Field_creation_commands(self):
+        if self.test_only and inspect.stack()[0][3] not in self.test_only:
+            return
+
+        f = self.f.copy()
+
+        for rd in (False, True):
+            for indent in (0, 4):
+                for h in (False, True):
+                    for s in (False, True):
+                        for ns in (None, ''):
+                            _ = f.creation_commands(
+                                representative_data=rd,
+                                indent=indent,
+                                namespace=ns,
+                                string=s,
+                                header=h)
+                            for i in range(7):
+                                f = cfdm.example_field(i)
+                                _ = f.creation_commands(
+                                    representative_data=rd,
+                                    indent=indent,
+                                    namespace=ns,
+                                    string=s,
+                                    header=h)
+        # --- End: for
+
 # --- End: class