+++ Changelog.rst
@@ -5,12 +5,16 @@ version 1.8.7.0
 **2020-09-??**
 
 * New keyword parameter to `cfdm.Field.set_data`: ``inplace``
+* New method: `cfdm.Field.creation_commands`
+* New method: `cfdm.Data.creation_commands`
 * New class: `cfdm.core.meta.DocstringRewriteMeta`
 * New method: `cfdm.Field._docstring_special_substitutions`
 * New method: `cfdm.Field._docstring_substitutions`
 * New method: `cfdm.Field._docstring_package_depth`
 * New method: `cfdm.Field._docstring_method_exclusions`
 * New function: `cfdm.unique_domains`
+* Enabled "creation commands" methods
+  (https://github.com/NCAS-CMS/cfdm/issues/53)
 * Fixed bug that caused failures when reading or writing a dataset
   that contains multiple geometry containers
   (https://github.com/NCAS-CMS/cfdm/issues/65)